Lesson 03 Homework Final Save
I finished replacing the ID-defined <div>'s in the HTML with
semantically correct tags.
I usually like to have my web browser sized to 50% of the width of my
screen, and noticed that when it is sized this way, the aside is moved
to the bottom of the page! Not very attractive.
I then discovered that by resizing the width value of the aside to less
than 23% allows the browser window to be significantly resized before
forcing to jump to the bottom of the page. The reason for this is not
arbitrary, but due to the padding defined in other elements within
<body>.