04LessonFour Checkpoint/Evaluation Save 4
I have cleaned up some css selectors which I will not be using, as the
modifiers are working in the IDs I have already created.
After lots of wrestling I found a solution online that allowed me to
align the two sections side by side.

The next thing left to do is define the READ MORE button width as being
seperate from its parent element, and then to fix the height of
section#rightbox so that it matches that of section#leftbox (width:
394px)