04LessonFour Evaluation/Checkpoint Save 5
I figured out how to fix the section#rightbox height. In the previous
version I was using padding definitions to vertically align the div
element within this section, but that was causing issues with the height
- the box was too tall.

Instead, I'm now defining "display: table" within the parent element
section#rightbox, then defining the child element div#vertalign with
"display:table-cell". This allows me to set "vertical-align: center"
within the child element and also maintain the proper height of the
section as a whole.

The only thing left to do is fix the READMORE button width.