04LessonFour Evaluation/Checkpoint Save 6
99% finished.
The READMORE button is now sized and centered appropriately. At first I
was trying to dial in the positioning by plugging in various values to
"margin-left", but I got the feeling this was crude / inexact. Instead,
since width and padding are already set, I used "margin: auto" to
automatically horizontal-center the READMORE element in its parent
section.

Tomorrow I'm going to review and clean up the code a little before fully
turning in the assignment.