04LessonFour Checkpoint/Evaluation Final Save
I took a minute to clean up a couple discrepancies in the code before
submitting my finished work.

In CSS there were a couple redundant definitions which I was able to
simplify - namely in the "body" and "section#leftbox" selectors, where I
had previously set values for "padding" and "margin" to "0px 0px 0px
0px". I have shortened these to simply "0px". Then, in
"section#rightbox" I was able to redefine "margin: 0px 0px 0px 3px" to
simply "margin-left: 3px". This particular change makes the code more
fluid in relation the way the styling affects its the respective visual
element - that is, it's easier as you're scanning the CSS to see how
this margin-left definition would be responsible for the small gap
between section#rightbox and section#leftbox.