In Class Project WIP
The basic structure of the website is in place, but there are various
problems at the moment. I did not use semantic elements, so the code is
kind of messy, and the text for the 3 articles on the right will change
position significantly if the width is changed slightly. I also need to
add borders between the elements in the lower section.