08LessonEight In-Class Save 2
Finished adding and aligning all elements in the page. I also managed to
figure out the source of the font used in the example image - Open Sans
- and embedded it via Google API.

I'm a little confused as to how I should reconcile the size of my page
with that of the example image, given the aforementioned discrepancies
in screen resolutions between my computer and that of the user who
created the example. Sounds like a good question to address in-class.