Winery Site Complete Feature and Brag Sections
Completed laying out and aligning the Feature and Brag sections. I used
background-position-x to scale the alignment of the rightside image in
the Feature section with a percentage. The image data is wholly defined
in css. The Brag section uses the same classes from the Splash section
to define the dimensions of the curtain and the title overlay. Applying
styling was very simple, just a matter of copying and pasting.