lesson 12
Lesson 12 all set up with css and Html. I tried to figure out how to
use the preset bootstrap like the sample did did fig caption and figure
but i did not figure it out.   I believe that with bootstrap you should
not use widths on css? is that correct? I did that with everything but
could not figure out way the image did not work the way it should.