Lesson Thirteen In-Class Commit
I managed to align all the elements with bootstrap as well as render
specific elements visible or hidden based on screen size. I am going to
forego fine-tuning the aesthetics of the page for the time being in so
that I can get started on the homework. I will also have to revisit
sizing the button in the news section - I haven't figured out how to
make it the same size as that used in the assignment's example image by
specifically using Bootstrap. Nevertheless it resizes and aligns
functionally in its current dimensions.