Checkpoint 3 WIP
Almost everything is done and works correctly, but I am still working on
the mobile view for #section_2. Instead of just having the image below
or above the text, I wanted to have it be the background because I felt
that too much space was being wasted by having a large image and minimal
text placed on top of each other and probably take up a large portion of
the screen. I have had a lot of trouble with this though. Currently, the
only problem is that the image does not resize properly at low screen
width values. I think another possible way would be to hide the actual
image and set the background image of the <div> that the text is in to
the image would work pretty well and make resizing very simple and
clean, and I will try to test this as soon as possible. There are also
issues with text being hard to read when in front of certain parts of
the image, but I can fairly easily fix this with either a stroke around
the text or, more preferably, an opacity that appears over the image on
mobile view. The only other concern I have is that the text is slightly
off from what it is on the original site, but again, all that requires
is finding the font on google fonts. I will try to have these things
added before class today.