Checkpoint 3
Instead of making the image appear behind the text as I was trying to do
before, I set the background to the image, making the css much simpler
and easier to manage. This also made the opacity very easy to apply. I
was also able to change the font to what it was on the site. Everything
seems to be working well, I see no current issues.