Lesson 16 Homework
Everything is mostly complete, I had trouble finding the right icons (I
checked on font-awesome and Google Fonts, but the closest icon I could
find to any on the page was a membership icon that somewhat resembled
the mail icon used on the page), so there are currently none included. I
added the animation by making all of the smaller boxes jiggle when
clicked on, however, there is an odd bug that I could not solve after
testing for awhile. When in the smallest view (and only the smallest
view), the first 2 small boxes will not jiggle when clicked on, although
all others after them will. Once you change to a larger view, however,
they work perfectly. I found that the function to add the class wasn't
being called at all, meaning that the onclick value wasn't being added.
However, I also had the console print all of the DOM elements that the
effects were being applied to, and they were included in the list. They
had also been given the animated class that was in the for loop that
also added the onclick value, so I do not see how they would not receive
it, especially when all the others did. I have a feeling that it has
something to do with the actual location of the div element being
located somewhere else from where the content actually is, or something
similar so that I am not clicking in the right location to trigger the
onclick event or something is preventing it from being called. Either
way, there are no other problems that I can currently find. The html and
css code is kind of sloppy and there may be a few unnecessary divs that
I can clean out, so I will make sure to do that soon.