Lesson 17 Homework
The site works relatively well, there are a few problems with
positioning that I had, but it isn't too bad. I had trouble with
relative and absolute positioning within bootstrap classes, so I had to
manually set a height, which also had its difficulties. Overall, it
could be a little better, but I think it works pretty well as it is. I
will see if I can improve it a little bit more before class.