Final Project
Mostly finished, I will try to comment the code for the "about" page as
I have tried to do for my other pages and javascript. Other than that,
there are no real problems that I have found, there may be some design
flaws with the site, but I think the code works well for the most part.
I do think the code could be more clean and efficient, and I may try to
improve it in the future, but for now, it seems to do what it needs to.