Final Project
After opening the project on my computer with a larger monitor, I
realized that my profile image in the "about" page was going off of the
div, so I set a max-height value. I also noticed that there is an issue
when scrolling where the "programming" section is skipped and not
highlighted when being scrolled by in the separate nav bar. I'm almost
certain that this is a result of the monitor size and having multiple
sections on the screen at the same time, because I have tried it at
other window heights. I could fix this by adding more content, but I
don't really having anything else to add, and I believe that this
problem will always exist if the monitor is large enough. Again, I used
some unconventional methods when coding parts of the page, so I will try
to find better ways of doing what I wanted to and try to change that as
soon as possible.