Final Project Final Changes
Created completely alternate site because I did not feel like the design
of my original site was very good. The new site does not have all of the
requirements of the original project, and there were many things I
wanted to implement that I was unable to due to time constraints, but I
intended it only as a side project to work on because I was finished
with the original.