Homework for Lesson 4
I ignored what you told us to do on the button and just messed around
with some css effects. I did the padding differently because I wanted
some space to work with. When the button does it’s effect it moves the
other text a bit which is strange and I want to figure out how to fix
it. It’s also not a smooth effect which I don’t like. I assume it has
to do with the text align: center; property. I need to work on getting
formatting correct especially moving text around divs and containers.