Intro Workbook Links
Completely re-styled my intro workbook links to fit more of my
personality/outlook. The pictures are NYC themed because my goal is to
move to the big city to be a web developer/freelance my work. So
looking at the pictures when I continue working on my assignments will
encourage me to strive to do my very best in everything I do to get
where I should be. This project will also be a great reference guide to
future assignments. Thanks for looking!