Lesson 8 | In Class Project
Linking is messed up. Need to fix. Second Pass through this assignment.
 Everything is starting to look good except for the aside. I need to
more time to figure out how to format this where within each div you
can see the image and the text to the right. also in the footer there
needs to be green check icons that I haven’t completed yet. work in
progress.