02Lesson\02Project\JasonCornish
For this webpage I made it a news article satire about the 2016 election
mocking both Hillary Clinton and Donald Trump. For the header I decided
to make the "logo" a picture of the forest that I found on google and
using CSS I merged the Daily Planet <h1>Header on top of the logo. I
added a main but only used the main to hold the section and article. For
<section> I listed the name of the news article and within <article> I
put the actual content of the news article. For <aside> I put a JW Chase
Morgan  advertisement which is another jab at Hillary Clinton because JW
Chase Morgan is one of Hillarys biggest campaign funders. For the footer
I used a simple copyright and year.