Lesson2Project1Final
Objection: The job is to build a website that reflects the services
offered. What he/she wants is an image at the top of the page (logo)
and a title with a paragraph with information about the company, a list
of the services offered, and a method of contacting the business. We
chose a Mobile Photo Service.