Lesson 2 - Homework 2 - Git Semantics
Objective:
There was a miscommunication in our team, the developer that was
working on the HTML part of the project created a website without
semantic elements, but the designer that worked on the CSS file used
semantic type selectors. Now the website doesn't look the way it
should. Please help us out!