03Lesson\01Project\index.html
For this webpage I copied the original by using semantic elements
instead of divs or spans, such as using a header and stylizing the
header to act the same as a div. For the images with text next to them I
used tables with one cell being an image and the other cell being the
text.