The test, yo
I used flex-direction: row to make the two images on the same row, then
from there I just used simple html to make the skeleton of the article
and filled it in with some CSS such as padding and text-align: center to
get it in the middle. To actually get the text in the center height-wise
I had to make an article inside of the stylized section, then I used
padding to push the text down. Pretty fun project!