03Lesson/02Project
I started fiddling around and changing code but then I realized that
this wasn't my original code and that I was changing someone else's
code. Not sure how I did that but this is what the result is. I'm not
sure if I should redo the entire project 02 becasue I wouldn't have used
tables to space out the text boxes and images on the page.