07 Checkpoint
I changed the theme of the webpage because of time constraints but I do
hope to in the future go back and make it look prettier, I just wanted
you to know that I understand the concepts. And for the css-reset I
totally forgot to do that :( bummer, but I understand what it does. For
accessing multiple pages I dont really understand what other pages you
want me to access because the whole webpage is based on one webpage, not
multiple. So I had my links send to different spots on the webpage
hopefully thats what you meant.