10Lesson/01Homework attempt 2
Got the bulk of it finished but still can't figure out how to move the
button to the right, tried margin: 0 auto; but that doesn't work. Would
using the Offset class work? I would have to put the button in it's own
row and I would think there would be a better way to make it center.
Also not sure how to change the background color of the
pricing/offerings lists